chain: Remove logs_in_block_range from EthereumAdapterTrait
